Load file upload JS for file upload question type regardless of LM enabled
