+++ assets/course-theme/learning-mode.js
@@ -6,7 +6,6 @@ import './featured-video-size';
 import './scroll-direction';
 import { initCompleteLessonTransition } from './complete-lesson-button';
 import { submitContactTeacher } from './contact-teacher';
-import { handleFileUpload } from './file-upload-question-type';
 import { toggleFocusMode } from './focus-mode';
 
 if ( ! window.sensei ) {
@@ -24,4 +23,3 @@ window.sensei.courseTheme = { toggleFocusMode, toggleSidebar };
 window.sensei.submitContactTeacher = submitContactTeacher;
 
 initCompleteLessonTransition();
-handleFileUpload();
