+++ templates/single-quiz/question-type-file-upload.php
@@ -14,6 +14,8 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+Sensei()->assets->enqueue( 'sensei-file-upload-question-type', 'js/file-upload-question-type.js', [],true );
+
 /**
  * Get the question data with the current quiz id
  * All data is loaded in this array to keep the template clean.
